エラー解消

+++ DirectXdesu/Engine/2D/Texture/KTexture.cpp
@@ -81,7 +81,7 @@ void KTexture::CreateTexture(const std::string& directoryPath, const std::string
 	D3D12_DESCRIPTOR_HEAP_DESC srvHeapDesc = {};
 	srvHeapDesc.Type = D3D12_DESCRIPTOR_HEAP_TYPE_CBV_SRV_UAV;
 	srvHeapDesc.Flags = D3D12_DESCRIPTOR_HEAP_FLAG_SHADER_VISIBLE; // シェーダーから見えるように
-	srvHeapDesc.NumDescriptors = kMaxSRVCount; // テクスチャの枚数
+	srvHeapDesc.NumDescriptors = static_cast<UINT>(kMaxSRVCount); // テクスチャの枚数
 	result = device->CreateDescriptorHeap(&srvHeapDesc, IID_PPV_ARGS(&srvHeap)); // 生成
 	assert(SUCCEEDED(result));
